+++ app/src/main/java/com/github/damontecres/wholphin/ui/discover/DiscoverPage.kt
@@ -41,6 +41,7 @@ fun DiscoverPage(
         listOf(
             stringResource(R.string.discover),
             stringResource(R.string.request),
+            stringResource(R.string.search),
         )
     var selectedTabIndex by rememberSaveable { mutableIntStateOf(rememberedTabIndex) }
     val tabFocusRequesters = remember(tabs) { List(tabs.size) { FocusRequester() } }
@@ -85,6 +86,16 @@ fun DiscoverPage(
             // Requests
             1 -> {
                 SeerrRequestsPage(
+                    focusRequesterOnEmpty = tabFocusRequesters.getOrNull(selectedTabIndex),
+                    modifier =
+                        Modifier
+                            .fillMaxSize(),
+                )
+            }
+
+            // Search
+            2 -> {
+                SeerrSearchPage(
                     focusRequesterOnEmpty = tabFocusRequesters.getOrNull(selectedTabIndex),
                     modifier =
                         Modifier